% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkColumnLabels.R
\docType{class}
\name{checkColumnLabelsResult-class}
\alias{checkColumnLabelsResult-class}
\alias{checkColumnLabelsResult}
\title{An S4 subclass of CheckResult for checkColumnLabels.}
\description{
An S4 subclass of CheckResult for checkColumnLabels.
}
\seealso{
\code{CheckResult}
}
