% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkColumnNameSyntax.R
\name{checkColumnNameSyntax}
\alias{checkColumnNameSyntax}
\title{Check data frame for proper column name syntax.
This includes: \cr 1) Containing alphanumeric characters and underscores
only. \cr 2) Starting with an alpha character only. \cr 3) All letters
being fully capitalized.}
\usage{
checkColumnNameSyntax(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
An S4 CheckResult object with contents:
}
\description{
Check data frame for proper column name syntax.
This includes: \cr 1) Containing alphanumeric characters and underscores
only. \cr 2) Starting with an alpha character only. \cr 3) All letters
being fully capitalized.
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of this check}

\item{\code{pass}}{TRUE if and only if all column names contain valid syntax}

\item{\code{msg}}{Character of details when check fails, otherwise empty}

\item{\code{data}}{List of
\itemize{
\item \code{cols_violated}: Column numbers which violate syntax rules
}}
}}

\author{
Jedid Ahn, Logan Lim
}
