% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkColumnNameSyntax.R
\docType{class}
\name{checkColumnNameSyntaxResult-class}
\alias{checkColumnNameSyntaxResult-class}
\alias{checkColumnNameSyntaxResult}
\title{An S4 subclass of CheckResult for checkColumnNameSyntax.}
\description{
An S4 subclass of CheckResult for checkColumnNameSyntax.
}
\seealso{
\code{CheckResult}
}
