% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDateFormat.R
\name{checkDateFormat}
\alias{checkDateFormat}
\title{Check whether an input character vector contains valid date format.
A valid date is the following:}
\usage{
checkDateFormat(dates)
}
\arguments{
\item{dates}{A character vector of dates.}
}
\value{
An S4 CheckResult object with contents:
}
\description{
YYYYMMMDD, valid years (4 digit number), valid months (JAN to DEC),
and valid days (1-30 or 1-31 depending on the month, 1-28 (for February),
and 1-29 (for February on leap years)).
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of this check}

\item{\code{pass}}{TRUE if and only if all dates are in valid format}

\item{\code{msg}}{Character of details when check fails, otherwise empty}

\item{\code{data}}{List of
\itemize{
\item \code{indices_violated}: Indices associated with date format violation
}}
}}

\author{
Logan Lim, Derek Beaton, Jedid Ahn
}
