% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDateFormat.R
\docType{class}
\name{checkDateFormatResults-class}
\alias{checkDateFormatResults-class}
\alias{checkDateFormatResults}
\title{An S4 subclass of CheckResult for checkDateFormat.}
\description{
An S4 subclass of CheckResult for checkDateFormat.
}
\seealso{
\code{CheckResult}
}
