% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDirNames.R
\name{checkDirNames}
\alias{checkDirNames}
\title{Check that directory names use alphanumerics and underscores, and make
sure first character is not a digit or underscore.}
\usage{
checkDirNames(dirNames)
}
\arguments{
\item{dirNames}{A character vector of names that we want to check.}
}
\value{
An S4 CheckResult object with contents:
}
\description{
Check that all strings are named strictly using alphanumeric characters
and underscores. In addition, check that the strings names do not start
with digits or underscores.
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of this check}

\item{\code{pass}}{TRUE if and only if all directory names contain valid syntax}

\item{\code{msg}}{Character of details when check fails, otherwise empty}

\item{\code{data}}{List of
\itemize{
\item \code{names_violated}: Names that failed the check as character vector
}}
}}

\author{
Jedid Ahn, Logan Lim
}
