% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDirNames.R
\docType{class}
\name{checkDirNamesResult-class}
\alias{checkDirNamesResult-class}
\alias{checkDirNamesResult}
\title{An S4 subclass of CheckResult for checkDirNames.}
\description{
An S4 subclass of CheckResult for checkDirNames.
}
\seealso{
\code{CheckResult}
}
