% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDirStructure.R
\name{checkDirStructure}
\alias{checkDirStructure}
\title{Check whether or not directory matches specified structure. Also checks a
specified set of names in a directory against the list.files output.}
\usage{
checkDirStructure(baseDir, dirStructure, recursive = T, exact = F)
}
\arguments{
\item{baseDir}{Base/top-most directory}

\item{dirStructure}{A character vector of file & directory names with
forward slashes.}

\item{recursive}{TRUE if you wish to parse into subdirectories, otherwise FALSE}

\item{exact}{When TRUE, the check will only pass if the files and directories
in the base folder match exactly. When FALSE, will pass when the base folder
contains all of the specified files and directories.}
}
\value{
An S4 CheckResult object with contents:
}
\description{
Check whether or not directory matches specified structure. Also checks a
specified set of names in a directory against the list.files output.
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of this check}

\item{\code{pass}}{TRUE if and only if baseDir matches specified structure as listed
in dirStructure}

\item{\code{msg}}{Character of details when check fails, otherwise empty}

\item{\code{data}}{List of
\itemize{
\item \code{fnames}: Vector of file and directory names in the base directory.
}}
}}

\author{
Logan Lim
}
