% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDirStructure.R
\docType{class}
\name{checkDirStructureResult-class}
\alias{checkDirStructureResult-class}
\alias{checkDirStructureResult}
\title{An S4 subclass of CheckResult for checkDirStructure.}
\description{
An S4 subclass of CheckResult for checkDirStructure.
}
\seealso{
\code{CheckResult}
}
