% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkEncoding.R
\name{checkEncoding}
\alias{checkEncoding}
\title{Check whether a tabular csv file is encoded in the broadly accepted formats
of ASCII or UTF-8; UTF-8 is preferred and suggested.}
\usage{
checkEncoding(filePath)
}
\arguments{
\item{filePath}{Tabular csv file path}
}
\value{
An S4 CheckResult object with contents:
}
\description{
Check whether a tabular csv file is encoded in the broadly accepted formats
of ASCII or UTF-8; UTF-8 is preferred and suggested.
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of this check}

\item{\code{pass}}{TRUE if and only if all encoding in file is valid}

\item{\code{msg}}{Character of details when check fails, otherwise empty}

\item{\code{data}}{List of
\itemize{
\item \code{encoding_probabilities}: Encoding probabilities of the file, if
encoding guess was performed.
\item \code{cols_w_no_ASCII}: Column numbers with ASCII violations.
}}
}}

\author{
Logan Lim, Derek Beaton, Jedid Ahn
}
