% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkEncoding.R
\docType{class}
\name{checkEncodingResult-class}
\alias{checkEncodingResult-class}
\alias{checkEncodingResult}
\title{An S4 subclass of CheckResult for checkEncoding.}
\description{
An S4 subclass of CheckResult for checkEncoding.
}
\seealso{
\code{CheckResult}
}
