% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkForBlankCells.R
\name{checkForBlankCells}
\alias{checkForBlankCells}
\title{Check for blank cells and NA in a data frame.}
\usage{
checkForBlankCells(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
An S4 CheckResult object with contents:
}
\description{
Check for blank cells and NA in a data frame.
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of this check}

\item{\code{pass}}{TRUE if and only if data frame contains no blank cells}

\item{\code{msg}}{Character of details when check fails, otherwise empty}

\item{\code{data}}{List of
\code{cols_violated}: column numbers containing a blank cell}
}}

\author{
Logan Lim, Jedid Ahn
}
