% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkForBlankCells.R
\docType{class}
\name{checkForBlankCellsResult-class}
\alias{checkForBlankCellsResult-class}
\alias{checkForBlankCellsResult}
\title{An S4 subclass of CheckResult for checkForBlankCells.}
\description{
An S4 subclass of CheckResult for checkForBlankCells.
}
\seealso{
\code{CheckResult}
}
