% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkForColumnNameDuplicates.R
\name{checkForColumnNameDuplicates}
\alias{checkForColumnNameDuplicates}
\title{Check data frame for duplicate column names (case insensitive).}
\usage{
checkForColumnNameDuplicates(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
An S4 CheckResult object with contents:
}
\description{
Check data frame for duplicate column names (case insensitive).
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of this check}

\item{\code{pass}}{TRUE if and only if data frame contains no duplicate column names}

\item{\code{msg}}{Character of details when check fails, otherwise empty}

\item{\code{data}}{List of
\itemize{
\item \code{duplicated_cols}: Duplicate column names
}}
}}

\author{
Logan Lim, Jedid Ahn
}
