% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkForColumnNameDuplicates.R
\docType{class}
\name{checkForColumnNameDuplicatesResult-class}
\alias{checkForColumnNameDuplicatesResult-class}
\alias{checkForColumnNameDuplicatesResult}
\title{An S4 subclass of CheckResult for checkForColumnNameDuplicates.}
\description{
An S4 subclass of CheckResult for checkForColumnNameDuplicates.
}
\seealso{
\code{CheckResult}
}
