% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkForCommas.R
\name{checkForCommas}
\alias{checkForCommas}
\title{Check data frame for commas.}
\usage{
checkForCommas(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
An S4 CheckResult object with contents:
}
\description{
Check data frame for commas.
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of this check}

\item{\code{pass}}{TRUE if and only if data frame contains no commas}

\item{\code{msg}}{Character of details when check fails, otherwise empty}

\item{\code{data}}{List of
\itemize{
\item \code{cols_violated}: column numbers associated with comma violation
}}
}}

\author{
Logan Lim, Jedid Ahn
}
