% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkForCommas.R
\docType{class}
\name{checkForCommasResult-class}
\alias{checkForCommasResult-class}
\alias{checkForCommasResult}
\title{An S4 subclass of CheckResult for checkForCommas.}
\description{
An S4 subclass of CheckResult for checkForCommas.
}
\seealso{
\code{CheckResult}
}
