% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkForCompressedFiles.R
\name{checkForCompressedFiles}
\alias{checkForCompressedFiles}
\title{Check if a directory contains compressed files.}
\usage{
checkForCompressedFiles(dirPath)
}
\arguments{
\item{dirPath}{Full directory path in character format}
}
\value{
An S4 CheckResult object with contents:
}
\description{
Check if a directory contains compressed files.
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of this check}

\item{\code{pass}}{TRUE if and only if directory contains no compressed files}

\item{\code{msg}}{Character of details when check fails, otherwise empty}

\item{\code{data}}{Empty list}
}}

\author{
Roberto Lentini, Logan Lim
}
