% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkForCompressedFiles.R
\docType{class}
\name{checkForCompressedFilesResult-class}
\alias{checkForCompressedFilesResult-class}
\alias{checkForCompressedFilesResult}
\title{An S4 subclass of CheckResult for checkForCompressedFiles.}
\description{
An S4 subclass of CheckResult for checkForCompressedFiles.
}
\seealso{
\code{CheckResult}
}
