% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkForDuplicateFileNames.R
\docType{class}
\name{checkForDuplicateFileNamesResult-class}
\alias{checkForDuplicateFileNamesResult-class}
\alias{checkForDuplicateFileNamesResult}
\title{An S4 subclass of CheckResult for checkForDuplicateFileNames.}
\description{
An S4 subclass of CheckResult for checkForDuplicateFileNames.
}
\seealso{
\code{CheckResult}
}
