% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkForEncapsulation.R
\name{checkForEncapsulation}
\alias{checkForEncapsulation}
\title{Check data frame for proper column name syntax.}
\usage{
checkForEncapsulation(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
An S4 CheckResult object with contents:
}
\description{
Check whether a data frame contains proper encapsulation, such
that all data cells DO NOT have the following:
\enumerate{
\item Double quotes on 1 side only.
\item Double quotes within a data cell.
\item No single quotes unless it is an apostrophe.
}
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of this check}

\item{\code{pass}}{TRUE if and only if data frame contains no quote violations}

\item{\code{msg}}{Character of details when check fails, otherwise empty}

\item{\code{data}}{List of
\itemize{
\item \code{cols_violated}: Column numbers associated with encapsulation violation
}}
}}

\author{
Jedid Ahn, Logan Lim
}
