% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkForEncapsulation.R
\docType{class}
\name{checkForEncapsulationResult-class}
\alias{checkForEncapsulationResult-class}
\alias{checkForEncapsulationResult}
\title{An S4 subclass of CheckResult for checkForEncapsulation.}
\description{
An S4 subclass of CheckResult for checkForEncapsulation.
}
\seealso{
\code{CheckResult}
}
