% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkForFileTypes.R
\name{checkForFileTypes}
\alias{checkForFileTypes}
\title{Check for specific file types in directories.}
\usage{
checkForFileTypes(fileExt, dirPath, recursive = F)
}
\arguments{
\item{fileExt}{Character vector of file types to look for}

\item{dirPath}{Full directory path}

\item{recursive}{TRUE if you wish to parse into subdirectories, otherwise FALSE}
}
\value{
An S4 CheckResult object with contents:
}
\description{
Check for specific file types in directories.
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of this check}

\item{\code{pass}}{TRUE if and only if file types in fileExt are found in directory}

\item{\code{msg}}{Character of details when check fails, otherwise empty}

\item{\code{data}}{List of
\itemize{
\item \code{matched_names}: File names that match file types given in fileExt
}}
}}

\author{
Jedid Ahn, Logan Lim
}
