% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkForFileTypes.R
\docType{class}
\name{checkForFileTypesResult-class}
\alias{checkForFileTypesResult-class}
\alias{checkForFileTypesResult}
\title{An S4 subclass of CheckResult for checkForFileTypes.}
\description{
An S4 subclass of CheckResult for checkForFileTypes.
}
\seealso{
\code{CheckResult}
}
