% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkForIDOverlap.R
\name{checkForIDOverlap}
\alias{checkForIDOverlap}
\title{Check for participant ID overlap between 2 data frames. In ONDRI case,
the participant IDs in the DATA and MISSING files should be checked for
potential overlap.}
\usage{
checkForIDOverlap(df1, df2)
}
\arguments{
\item{df1}{Data frame 1, such as a DATA data frame}

\item{df2}{Data frame 2, such as a MISSING data frame}
}
\value{
An S4 CheckResult object containing:
}
\description{
Check for participant ID overlap between 2 data frames. In ONDRI case,
the participant IDs in the DATA and MISSING files should be checked for
potential overlap.
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of this check}

\item{\code{pass}}{TRUE if and only if no participant IDs overlap}

\item{\code{msg}}{Character of details when check fails, otherwise empty}

\item{\code{data}}{List of
\itemize{
\item \code{overlapping_ids}: Vector of participant IDs that overlap.
}}
}}

\author{
Jedid Ahn
}
