% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkForIDOverlap.R
\docType{class}
\name{checkForIDOverlapResults-class}
\alias{checkForIDOverlapResults-class}
\alias{checkForIDOverlapResults}
\title{An S4 subclass of CheckResult for checkForIDOverlap.}
\description{
An S4 subclass of CheckResult for checkForIDOverlap.
}
\seealso{
\code{CheckResult}
}
