% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkForRaggedRows.R
\name{checkForRaggedRows}
\alias{checkForRaggedRows}
\title{Check whether a tabular csv file contains ragged rows.}
\usage{
checkForRaggedRows(filePath)
}
\arguments{
\item{filePath}{Tabular csv file path}
}
\value{
An S4 CheckResult object containing:
}
\description{
Check whether a tabular csv file contains ragged rows.
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of this check}

\item{\code{pass}}{TRUE if and only if file contains no ragged rows}

\item{\code{msg}}{Character of details when check fails, otherwise empty}

\item{\code{data}}{Empty list (for now)}
}}

\author{
Logan Lim, Derek Beaton, Jedid Ahn
}
