% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkForRaggedRows.R
\docType{class}
\name{checkForRaggedRowsResult-class}
\alias{checkForRaggedRowsResult-class}
\alias{checkForRaggedRowsResult}
\title{An S4 subclass of CheckResult for checkForRaggedRows.}
\description{
An S4 subclass of CheckResult for checkForRaggedRows.
}
\seealso{
\code{CheckResult}
}
