% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkForSpecialCharacters.R
\name{checkForSpecialCharacters}
\alias{checkForSpecialCharacters}
\title{Check for special characters in a data frame.}
\usage{
checkForSpecialCharacters(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
An S4 CheckResult object containing:
}
\description{
Check for special characters in a data frame.
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of this check}

\item{\code{pass}}{TRUE if and only if data frame contains no special characters}

\item{\code{msg}}{Character of details when check fails, otherwise empty}

\item{\code{data}}{List of
\itemize{
\item \code{cols_violated}: Column numbers associated with special character violation
}}
}}

\author{
Logan Lim, Derek Beaton, Jedid Ahn
}
