% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkForSpecialCharacters.R
\docType{class}
\name{checkForSpecialCharactersResult-class}
\alias{checkForSpecialCharactersResult-class}
\alias{checkForSpecialCharactersResult}
\title{An S4 subclass of CheckResult for checkForSpecialCharacters.}
\description{
An S4 subclass of CheckResult for checkForSpecialCharacters.
}
\seealso{
\code{CheckResult}
}
