% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkForSpecialValues.R
\docType{class}
\name{checkForSpecialValues-class}
\alias{checkForSpecialValues-class}
\alias{checkForSpecialValuesResult}
\title{An S4 subclass of CheckResult for checkForSpecialValues.}
\description{
An S4 subclass of CheckResult for checkForSpecialValues.
}
\seealso{
\code{CheckResult}
}
