% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkForSpecialValues.R
\name{checkForSpecialValues}
\alias{checkForSpecialValues}
\title{Check whether a tabular csv file contains special values such as NaN
(not a number), Inf (infinity), -Inf (negative infinity), or NULL (no value).}
\usage{
checkForSpecialValues(filePath)
}
\arguments{
\item{filePath}{Tabular csv file path}
}
\value{
An S4 CheckResult object containing:
}
\description{
Check whether a tabular csv file contains special values such as NaN
(not a number), Inf (infinity), -Inf (negative infinity), or NULL (no value).
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of this check}

\item{\code{pass}}{TRUE if and only if file contains no whitespaces}

\item{\code{msg}}{Character of details when check fails, otherwise empty}

\item{\code{data}}{List of
\itemize{
\item \code{cols_violated}: Column numbers associated with special values violation
}}
}}

\author{
Logan Lim, Jedid Ahn
}
