% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkForWhiteSpaces.R
\name{checkForWhiteSpaces}
\alias{checkForWhiteSpaces}
\title{Check for leading or trailing whitespaces in a data frame.}
\usage{
checkForWhiteSpaces(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
An S4 CheckResult object containing:
}
\description{
Check for leading or trailing whitespaces in a data frame.
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of this check}

\item{\code{pass}}{TRUE if and only if data frame contains no whitespaces}

\item{\code{msg}}{Character of details when check fails, otherwise empty}

\item{\code{data}}{List of
\itemize{
\item \code{cols_violated}: Column numbers associated with whitespace violation
}}
}}

\author{
Logan Lim, Jedid Ahn
}
