% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkForWhiteSpaces.R
\docType{class}
\name{checkForWhiteSpacesResult-class}
\alias{checkForWhiteSpacesResult-class}
\alias{checkForWhiteSpacesResult}
\title{An S4 subclass of CheckResult for checkForWhiteSpaces.}
\description{
An S4 subclass of CheckResult for checkForWhiteSpaces.
}
\seealso{
\code{CheckResult}
}
