% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkM_Pattern.R
\name{checkM_Pattern}
\alias{checkM_Pattern}
\title{Check the ^M_ pattern in a data frame and count the number of occurrences
of each unique code.}
\usage{
checkM_Pattern(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
An S4 CheckResult object containing:
}
\description{
Check the ^M_ pattern in a data frame and count the number of occurrences
of each unique code.
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of this check}

\item{\code{pass}}{TRUE if and only if data frame contains an M_ pattern}

\item{\code{msg}}{Character of details of the pattern search}

\item{\code{data}}{List of
\itemize{
\item \code{unique_codes}: Vector outlining all unique codes with an ^M_ pattern.
\item \code{tally}: Table outlining the number of occurrences of each unique code.
}}
}}

\author{
Logan Lim, Derek Beaton, Jedid Ahn
}
