% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkM_Pattern.R
\docType{class}
\name{checkM_PatternResult-class}
\alias{checkM_PatternResult-class}
\alias{checkM_PatternResult}
\title{An S4 subclass of CheckResult for checkM_Pattern.}
\description{
An S4 subclass of CheckResult for checkM_Pattern.
}
\seealso{
\code{CheckResult}
}
