% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkMissingCodes.R
\name{checkMissingCodes}
\alias{checkMissingCodes}
\title{Check whether a vector contains valid missing codes wherever a
cell starts with "M_".}
\usage{
checkMissingCodes(
  dataVec,
  validMissingCodes = c("M_CB", "M_PI", "M_VR", "M_AE", "M_DNA", "M_TE", "M_NP",
    "M_ART", "M_TBC", "M_OTHER")
)
}
\arguments{
\item{dataVec}{A vector of data, such as a row or a column in a data frame.}

\item{validMissingCodes}{An optional vector of valid missing codes, with
the default value representing ONDRI missing codes.}
}
\value{
An S4 CheckResult object containing:
}
\description{
Check whether a vector contains valid missing codes wherever a
cell starts with "M_".
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of this check}

\item{\code{pass}}{TRUE if and only if vector contains valid missing codes}

\item{\code{msg}}{Character of details when check fails, otherwise empty}

\item{\code{data}}{List of
\itemize{
\item \code{indices_violated}: Indices associated with missing code violation
}}
}}

\author{
Logan Lim, Derek Beaton, Jedid Ahn
}
