% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkMissingCodes.R
\docType{class}
\name{checkMissingCodesResult-class}
\alias{checkMissingCodesResult-class}
\alias{checkMissingCodesResult}
\title{An S4 subclass of CheckResult for checkMissingCodes.}
\description{
An S4 subclass of CheckResult for checkMissingCodes.
}
\seealso{
\code{CheckResult}
}
