% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkNumOfCharacters.R
\name{checkNumOfCharacters}
\alias{checkNumOfCharacters}
\title{Check whether a data frame contains less than k characters in each cell.}
\usage{
checkNumOfCharacters(df, k = 200)
}
\arguments{
\item{df}{A data frame}

\item{k}{Optional parameter signifying the max number of characters allowed.
Default is 200 characters.}
}
\value{
An S4 CheckResult object containing:
}
\description{
Check whether a data frame contains less than k characters in each cell.
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of this check}

\item{\code{pass}}{TRUE if and only if data frame contains less than k characters}

\item{\code{msg}}{Character of details when check fails, otherwise empty}

\item{\code{data}}{List of
\itemize{
\item \code{cols_violated}: Column numbers associated with number of characters violation
}}
}}

\author{
Jedid Ahn, Logan Lim
}
