% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkNumOfCharacters.R
\docType{class}
\name{checkNumOfCharactersResult-class}
\alias{checkNumOfCharactersResult-class}
\alias{checkNumOfCharactersResult}
\title{An S4 subclass of CheckResult for checkNumOfCharacters.}
\description{
An S4 subclass of CheckResult for checkNumOfCharacters.
}
\seealso{
\code{CheckResult}
}
