% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckResult.R
\docType{class}
\name{CheckResult-class}
\alias{CheckResult-class}
\alias{CheckResult}
\title{The base class for the results of a check function.}
\description{
The base class for the results of a check function.
}
\section{Slots}{

\describe{
\item{\code{name}}{\emph{character.} Name of the check function that returned this object.}

\item{\code{pass}}{\emph{logical.} Whether or not the check passed/failed.}

\item{\code{msg}}{\emph{character.} A message associated with pass/fail value. Gives more
information after the check.}

\item{\code{data}}{\emph{list.} Any assorted data returned from the check.}
}}

