% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckResult.R
\name{combineChecks}
\alias{combineChecks}
\title{'Concatenate' CheckResult objects for compound checks.}
\usage{
combineChecks(...)
}
\arguments{
\item{...}{CheckResult objects}
}
\value{
New CheckResult objects with slots that are a concatenation of
slots of multiple CheckResult objects.
If one CheckResult object is given, then it is given back.
If two or more are given, their \code{name}, \code{pass}, \code{msg}, \code{data} slots are
all combined into single vectors on one returned CheckResult object
}
\description{
'Concatenate' CheckResult objects for compound checks.
}
