% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareDICTColumnLabels.R
\name{compareDICTColumnLabels}
\alias{compareDICTColumnLabels}
\title{Compare column labels in the DICT data frame and the column names in the
DATA data frame.}
\usage{
compareDICTColumnLabels(dictDF, dataDF)
}
\arguments{
\item{dictDF}{Dictionary data frame}

\item{dataDF}{Data data frame}
}
\value{
An S4 CheckResult object containing:
}
\description{
Compare column labels in the DICT data frame and the column names in the
DATA data frame.
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of this check}

\item{\code{pass}}{TRUE if and only if column labels and column names match}

\item{\code{msg}}{Character of details when check fails, otherwise empty}

\item{\code{data}}{List of
\itemize{
\item \code{missing_from_data}: Column labels in the DICT file that are not in the DATA file
\item \code{missing_from_dict}: Column names in the DATA file that are not in the DICT file
}}
}}

\author{
Jedid Ahn, Logan Lim
}
