% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareDICTColumnLabels.R
\docType{class}
\name{compareDICTColumnLabelsResult-class}
\alias{compareDICTColumnLabelsResult-class}
\alias{compareDICTColumnLabelsResult}
\title{An S4 subclass of CheckResult for compareDICTColumnLabels.}
\description{
An S4 subclass of CheckResult for compareDICTColumnLabels.
}
\seealso{
\code{CheckResult}
}
