% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareParticipantFileNames.R
\name{compareParticipantFileNames}
\alias{compareParticipantFileNames}
\title{Compare a vector of file names (such as the FILENAME column in the level 3
FILELIST file) and the literal contents of a directory (such as the level 3
DATAFILES folder) and check whether they match exactly.}
\usage{
compareParticipantFileNames(fileNamesVec, dirPath)
}
\arguments{
\item{fileNamesVec}{Character vector of participant file names}

\item{dirPath}{Directory path of participant files}
}
\value{
An S4 CheckResult object containing:
}
\description{
Compare a vector of file names (such as the FILENAME column in the level 3
FILELIST file) and the literal contents of a directory (such as the level 3
DATAFILES folder) and check whether they match exactly.
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of this check}

\item{\code{pass}}{TRUE if and only if all participant files exist in vector
of participant file names}

\item{\code{msg}}{Character of details when check fails, otherwise empty}

\item{\code{data}}{List of
\itemize{
\item \code{missing_from_vec}: File names in the participant files folder that are
not in the vector of participant file names.
\item \code{missing_from_dir}: File names in the vector of participant file names
that are not in the participant files folder.
}}
}}

\author{
Jedid Ahn, Jeremy Tanuan
}
