% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareParticipantFileNames.R
\docType{class}
\name{compareParticipantFiles-class}
\alias{compareParticipantFiles-class}
\alias{compareParticipantFileNameResults}
\title{An S4 subclass of CheckResult for compareParticipantFileNames.}
\description{
An S4 subclass of CheckResult for compareParticipantFileNames.
}
\seealso{
\code{CheckResult}
}
