% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareREADMEFileNames.R
\docType{class}
\name{compareREADMEFileNames-class}
\alias{compareREADMEFileNames-class}
\alias{compareREADMEFileNamesResult}
\title{An S4 subclass of CheckResult for compareREADMEFileNames.}
\description{
An S4 subclass of CheckResult for compareREADMEFileNames.
}
\seealso{
\code{CheckResult}
}
