% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareREADMEFileNames.R
\name{compareREADMEFileNames}
\alias{compareREADMEFileNames}
\title{Compare the file names in the README file and the files in the directory
containing the README file.}
\usage{
compareREADMEFileNames(readmeDF, dirPath)
}
\arguments{
\item{readmeDF}{README data frame}

\item{dirPath}{Full directory path containing data package files}
}
\value{
An S4 CheckResult object containing:
}
\description{
Compare the file names in the README file and the files in the directory
containing the README file.
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of this check}

\item{\code{pass}}{TRUE if and only if file names and directory files match}

\item{\code{msg}}{Character of details when check fails, otherwise empty}

\item{\code{data}}{List of
\itemize{
\item \code{missing_from_dir}: File names in the directory that are not in the README file
\item \code{missing_from_csv}: File names in the README file that are not in the directory
}}
}}

\author{
Jedid Ahn, Logan Lim, Jeremy Tanuan
}
