% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertDataFrame.R
\name{convertDataFrame}
\alias{convertDataFrame}
\title{Convert the initial data frame to make values consistent to allow for
easier analyzation.}
\usage{
convertDataFrame(df)
}
\arguments{
\item{df}{Initial data frame}
}
\value{
df Converted data frame
}
\description{
Convert the initial data frame to make values consistent to allow for
easier analyzation.
}
\author{
Jedid Ahn, Logan Lim
}
