% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareParticipantFileNames.R
\name{getFileName}
\alias{getFileName}
\title{HELPER FUNCTION: Return the file name of a tabular file depending on
its type.}
\usage{
getFileName(dirPath, fileType)
}
\arguments{
\item{dirPath}{Directory path}

\item{fileType}{File type such as DATA, DICT, README, or MISSING}
}
\value{
Full file name of file type
}
\description{
HELPER FUNCTION: Return the file name of a tabular file depending on
its type.
}
\author{
Jedid Ahn
}
