% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckResult.R
\name{stopIfFail}
\alias{stopIfFail}
\title{Throw an error if check fails}
\usage{
stopIfFail(check.res)
}
\arguments{
\item{check.res}{Check function

If CheckResult@pass is FALSE, then throw an error with message =
CheckResult@msg If check.res is an object from \link{combineChecks} then an error
will be thrown if any of them fail. Also throws error if object is not of
type CheckResult. Otherwise, return check.res.}
}
\description{
Throw an error if check fails
}
